Deleted the factory classes, and their test classes. I misunderstood the point of a factory, and just transfered the methods they used to board. This makes sense because board is figuring out which tiles are which, and which pieces are which, and then populating its own data structures. It also cuts down on how many classes we have. I moved the tests in the factoryTest classes to board. Everything still passes just fine :)
